% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_MCSim_beta.R
\name{pi_MCSim_beta}
\alias{pi_MCSim_beta}
\title{Beta Prior Simulation for Binomial RDT}
\usage{
pi_MCSim_beta(M, seed, a, b)
}
\arguments{
\item{M}{Simulation sample size}

\item{seed}{Random seed for random sample}

\item{a}{Shape parameter 1 for beta distribution}

\item{b}{Shape parameter 2 for beta distribution}
}
\value{
Vector of failure probability sample values
}
\description{
Define the simulation function to generate failure probability with Beta prior distributions as conjugate prior to binomial distributions (for binomial RDT).
}
\examples{
pi <- pi_MCSim_beta(M = 1000, seed = 10, a = 1, b = 1)
}
\seealso{
\code{\link{pi_MCSim_dirichlet}}

Other Prior distribution generation functions: 
\code{\link{pi_MCSim_dirichlet}()}
}
\concept{Prior distribution generation functions}
