% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_MCSim_dirichlet.R
\name{pi_MCSim_dirichlet}
\alias{pi_MCSim_dirichlet}
\title{Dirichlet Prior Simulation for Multi-state RDT}
\usage{
pi_MCSim_dirichlet(M, seed, par)
}
\arguments{
\item{M}{Simulation sample size}

\item{seed}{Random seed for random sample}

\item{par}{Parameters for dirichlet distribution}
}
\value{
Vector of failure probability sample
}
\description{
Define the simulation function to generate failure probability with Dirichlet prior distributions as conjugate prior to multinomial distributions (for multi-state RDT).
}
\examples{
pi <- pi_MCSim_dirichlet(M = 1000, seed = 10, par = c(1, 1, 1))
}
\seealso{
\code{\link{pi_MCSim_beta}}

Other Prior distribution generation functions: 
\code{\link{pi_MCSim_beta}()}
}
\concept{Prior distribution generation functions}
